ProductDisplay component renamed along with children
