+++ frontend/src/components/Products/ProductsRating.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types'
 
-const Rating = ({value, text, color}) => {
+const ProductsRating = ({value, text, color}) => {
     return (
         <div className='rating'>
         <span>
@@ -25,14 +25,14 @@ const Rating = ({value, text, color}) => {
     )
 };
 
-Rating.defaultProps = {
+ProductsRating.defaultProps = {
     color: '#f8e825',
 };
 
-Rating.propTypes = {
+ProductsRating.propTypes = {
     value: PropTypes.number.isRequired,
     text: PropTypes.string.isRequired,
     color: PropTypes.string,
 };
 
-export default Rating
+export default ProductsRating
